comment out testing lines

+++ wechat/assets/te_1843/img.py
@@ -6,6 +6,7 @@
 # `telongreads.py` file
 
 import requests
+from pathlib import Path
 from html.parser import HTMLParser
 from urllib.request import urlopen
 
@@ -44,12 +45,14 @@ class GetImage(HTMLParser):
         with requests.Session() as s: 
             for i in range(len(self.img_srcs)):
                 r = s.get(self.img_srcs[i]) 
-                if i == 0:
+                if i == 0 and Path(path).exists():
                     #TODO
-                    # use os to make sure img exists or create it if it doesn't
+                    # use os to make sure img exists 
                     with open(f'{path}/header-img.{self.img_srcs[i][-3:]}', 'wb') as img: 
                         img.write(r.content)
                 else:
+                    # or create it if it doesn't
+                    Path('img').mkdir()
                     with open(f'{path}/body-{i}.{self.img_srcs[i][-3:]}', 'wb') as img:
                             img.write(r.content)
